REFACTOR: GRAHPS
