+++ client/src/App.js
@@ -1,7 +1,6 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Home from './pages/home';
 import Predictor from './pages/predictor';
-import Graphs from './pages/RegressionGraph';
 import Research from './pages/research';
 
 function App() {
