+++ client/src/components/graphs/RegressionGraph.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import Plot from "react-plotly.js";
-import CompassLoader from "../components/compassLoader";
+import CompassLoader from "../compassLoader";
 
 export default function RegressionGraph({date, hour, stationNumber, width, height}) {
   let [points, setPoints] = useState([]);
